added new sentence with dialog

+++ lib/cutara/tasks/build.rake
@@ -9,7 +9,7 @@ namespace "cutara" do
   task :build, [:project, :testcase, :execution] => :download do
     mkdir_p(Cutara::PAGES) unless File.exists?(Cutara::PAGES)
     mkdir_p(Cutara::STEPS) unless File.exists?(Cutara::STEPS)
-    cp "#{Cutara::ASSETS}/env.rb", Cutara::SUPPORT
+    cp "#{Cutara::ASSETS}/env.rb", Cutara::SUPPORT unless File.exists? "#{Cutara::SUPPORT}/env.rb"
     steps_source = Cutara::ASSETS + "/step_definitions/" + @lang
     Dir.glob("#{steps_source}/*") {|f| cp File.expand_path(f), Cutara::STEPS }
   end
